+++ CHANGELOG.md
@@ -1,21 +1,35 @@
 # Changelog
 
-## [2.0.0] - 2019-10-18
+## [2.1.0] - 2019-12-20
+
 ### Added
+
+- `verifyChallenge` exported from index
+
+### Changed
+
+- code/comment formatting
+- refactored `random` function
+
+## [2.0.0] - 2019-10-18
+
+### Added
+
 - CHANGELOG
 - typescript definition
 - Cryptographically secured method for generating code verifier
 - Method for base64 url encoding
 
 ### Removed
+
 - `generateVerifier` export from index
 - `generateChallenge` export from index
 - `base64url` npm dependency
 - `randomatic` npm dependency
 
 ### Contributors
+
 - [lordnox]
 
 [lordnox]: https://github.com/lordnox
-
 [2.0.0]: https://github.com/crouchcd/pkce-challenge/releases/tag/2.0.0